classification-backend/requirements.txt
Flask==3.1.2
numpy==2.2.6
opencv_python==4.12.0.88
Pillow==12.0.0
boto3==1.40.59
dotenv==0.9.9
requests==2.32.5
python-dotenv==1.2.1
duckdb